Initialize LUA ref objects when growing the pool
This fixes saving/loading for me.
The problem was that when unitialized objects were saved, their type
and value could be anything. On loading, LUA_T_USERDATA value
pointers are searched for in a global array, and said random
unitialized values are, of course, not in that array, so loading such
a save triggered an assert().

Interestingly, it only failed with a 64bit build for me, the 32bit
build works fine regardless.