Fix sign error in Costume::moveHead so character looks down properly, instead of looking up when it should be looking down.
Fixes the clown guy to not tilt its head sideways. Note that this does not still quite reproduce the clown head motion of the original game. It seems to use a different mechanism for clamping the head yaw/pitch/roll max angles, and hence the clown looks down in a slightly different angle.