GRIM/EMI: Re-use TextObjects with the same textID and layer
Re-use existing TextObjects in MakeTextObjects() in order
to avoid a massive duplications in case the LUA code
calls MakeTextObject multiple times for the same text
but with a different color.

Note:
In EMI, the menus are using 2 TextObjects with the same textID
with different colors but with a different layer.

This fixes the issue that in the credits of EMI the colors
were off since there were multiple TextObjects for the same
textID but each with a different color (one per fade step).