TINYGL: Ignore zero-coordinate textures. XXX: incomplete
The holepunch in ha has what is likely a data error: all textures
coordinates are (0, 0). TinyGL used to render the holepunch in the (likely
intended) dark blue color from the texture. But this is inconsistently
handled between TinyGL and OpenGL (with and without shaders): the latter
ignore the texture, and render the holepunch as shaded grey. So make
TinyGL behave consistently - data error can be fixed independently.

Original Grim software renderer display the holepunch in a very dark grey
color, making it hard to tell if it takes the texture into account at all.
Data remained the same in remastered edition, which displays the holepunch
as light grey (on linux at least), consistently with residual OpenGL
renderer.

The grim reaper costume in lr also has several faces with the same
coordinates - just not zero X and Y - and they are textured with OpenGL,
so testing for all coordinates being equal is incorrect.

Also, glottis' fingers do have zero X and Y coordinates for all vertices
but they are still properly rendered in OpenGL... So this fix isn't quite
right either.