Added The Longest Journey
There are some *TODO* spots to fill yet as I have to find a retail copy of the game and I didn't try debugging commands. I have yet to complete the game on ResidualVM as well to confirm it's completable (with or without glitches, though some things are yet to be implemented, namely shadows and parts of UI). I wasn't sure if digital distribution versions (Steam and GOG) warrant a separate section and I have no idea if you can use iOS version of TLJ with ResidualVM. Keyboard shortcuts mentions in manual doesn't work yet and TLJ should be mentioned in http://www.residualvm.org/ as well, but I thought it's a good idea to let people know it's supported without looking down at code/commits.