MYST3: Add Saving from GMM
The Menu object internally manage a thumbnail
for saved games. But now a function has been
added to Menu to have a new thumbnail generated.

A saveGame function has been added to Myst3 that
just grabs this already generated thumbnail.
Only if the thumbnail is valid will the game
be allowed to save.

When saving using the GMM if no name is given
the save name will default to timestamps
with the "/" removed.

For saving in the original menu the default saved
name of "Saved Game" will be used when a user has
no specified a name. This is what the original does.