STARK: Bugfix for my previous PR
I accidentally forgot to change the name of the function to match, so the program won't compile now. This should be fixed with that PR, sorry!
I have to admit that I am only testing this change now (sorry again! but everything seems to work on my side now), but I'm surprised the checks went through.