MYST3: Don't repeat first sin value in Effect
Two of the sin loops evaluate sin(2 pi 255/255) at the
end of the sine loop which is the same as the first index,
sin(2 pi * 0), so when update() is called twice back to back
the last value of the first call, sin(2 pi 255/255) and the
first value of the next value sin(2 pi * 0) will be the same
so in effect it will be repeating the same sine value twice.

This changes the loop so it indexes the sin with 256 pieces
so that the last value is sin(2 pi * 255/256).