GRIM: Fix heap-use-after-free reordering deletion of pools
Found in GRIMDEMO: leave the underground office with the lift, exit the street and quit the game.

Exiting the game here deallocates object pools. That deletes the Costume object that resets AnimComponent,
that itself call setKey() but Set object has just been deleted previously.

The fix consists in reordering the deletions in clearPools() in order to avoid this kind of dependency.