EMI: Check _musicTable is null before it is used
After having installed EMI data and started the game, if we are in the situation
where it is asked to rename resource file FullMonkeyMap.imt from both CDs
into FullMonkeyMap1.imt and FullMonkeyMap2.imt, the game tries to start but
leads to an immediate segmentation fault.

In EMISound::setMusicState, _musicTable is checked againt nullptr but ... after
it is used. This check is moved up.