CONFIGURE: Add an environemt variable that discards the --host option
In the previous commit we modified the configure script to silently
ignore some options, that are not needed for ScummVM, but are valid
for the configure script generated by GNU Autotools, so the configure
macro for RPM-based distributions can be used when building a rpm
package for them.

Unfortunately the configure script, as used by ScummVM uses the --host
option to determine the target system it will actually be build for.
Autotools based configure scripts have a --target option for such a
purpose, and use the --host option to determine the system the build
is performed on.

For that reason there should be a way to discard the parameters passed
to the configure script with the --host option.  The easiest approach
to achieve this goal, is to have an environment variable, which when
set influences the configure script to ignore the parameters of the
--host option.

Thus we introduced a variable called 'CONFIGURE_NO_HOST', that will
when set to anything, but an empty value, before invoking the configure
script, have the parameters of the --host option take no influence on
the configure stage (and the build stage as well).

Backported from scummvm/scummvm@8c28a2da