ALL: Add error checking to PEResources::getResource
For some reason the `game.exe` that comes with the steam version of TLJ
is truncated. It is only 95744 bytes and contains a resource table but
not the resource data.

This currently results in random data from the binary to be passed into
the BMP parser in `DialogBox::loadBackground()`, resulting in a warning
`Only Windows v3 & v4 bitmaps are supported`. It could be worse, but
it sets developers on the wrong trail toward trying to fix BMP
parsing.

Avoid doing this by adding a check that both the seek and the read succeed,
which if so, makes loading the resource simply fail.